test: now testing import and use of common libraries
